Add in Jeff Gilfelt's "mapview balloon overlay" android library project.
We're packaging up the source, but even the standard build environment
reads library project source files when building an app that uses the
library.  The kludgery here is actually worse; I've got aapt generating
a single R.java file for everything, and there's a hand-hacked set of
stubs in the "mapviewballoons" directory which copies the ids for the
resources that the library cares about.

Sigh...