First cut at proximity alerts for shops.
Has problems --- status bar icon is just a random thing I had lying
around, and updates are kludgery.

Still, at this point we have the basics of what an ORM really needs
to support cleanly:  simple CRUD, basic hasMany/belongsTo associations,
and manipulation of managed objects both from the UI (which wants DB
action on background threads) and in the background (specifically, a
BroadcastReceiver for ProximityAlerts which sets up notifications).